+++ src/changelog.js
@@ -3,7 +3,7 @@ const readFile = util.promisify(require("fs").readFile);
 
 const shiftTillEntry = (lines, start) => {
   let line = lines[0];
-  while ((line !== undefined) && !line.startsWith(start)) {
+  while (line !== undefined && !line.startsWith(start)) {
     lines.shift();
     line = lines[0];
   }
@@ -11,7 +11,7 @@ const shiftTillEntry = (lines, start) => {
 
 const outputTillEntry = (lines, start) => {
   let line = lines[0];
-  while ((line !== undefined) && !line.startsWith(start)) {
+  while (line !== undefined && !line.startsWith(start)) {
     console.log(line);
     lines.shift();
     line = lines[0];
@@ -19,20 +19,22 @@ const outputTillEntry = (lines, start) => {
 };
 
 const outputChangelog = async config => {
-  const { diffTag, reportTag } = config;
+  const { diffTag, reportTag, releaseTag, releaseDiffTag } = config;
 
   try {
     const log = await readFile("CHANGELOG.md", "utf8");
     const lines = log.split("\n");
 
     console.log(`## Changelog since ${diffTag || "start"}\n`);
-    const start = reportTag === "HEAD" ? "##" : `## [${reportTag}]`;
+    const start = reportTag === "HEAD" ? "##" : `## [${releaseTag}]`;
 
     shiftTillEntry(lines, start);
-    outputTillEntry(lines, `## [${diffTag}]`);
+    outputTillEntry(lines, `## [${releaseDiffTag}]`);
   } catch (e) {
-    console.log("## No Changelog found\n")
-    console.log("Please add a CHANGELOG.md according to <https://keepachangelog.com/>")
+    console.log("## No Changelog found\n");
+    console.log(
+      "Please add a CHANGELOG.md according to <https://keepachangelog.com/>"
+    );
   }
 
   console.log("");
@@ -40,4 +42,4 @@ const outputChangelog = async config => {
 
 module.exports = {
   outputChangelog
-}
+};