+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+### Added
+- Support for dates as `--from` and `--till` options
+- Alias `--since` for `--from` and `--until` for `--till`
+- Extended contributor table with additions and deletions per author
+
 ## [1.1.0] - 2018-10-01
 ### Added
 - Support for `--from` and `--till` options